CMPE487/Final Project/requirements.txt
Click==7.0
tabulate==0.8.6
attrs==19.3.0
cattrs==0.9.0
